Implement job-based crawling and enhance results format
- Add crawl_jobs table for tracking individual crawl jobs
- Update crawls table with job_id, content_hash, and timestamp fields
- Implement content change detection using MD5 hashing
- Modify /crawl endpoint to return job_id
- Update /results endpoint to filter by job_id and improve result format
- Add last_crawled and last_modified fields to track page updates
- Optimize crawler logic for better performance and accuracy
- Update documentation and project summary
